requirements.txt
Pillow
exif
folium
streamlit
streamlit_folium
